move common generate code out into GenerateAction and GenerateParams
